requirements.txt
# Core dependencies
requests>=2.31.0
websocket-client>=1.6.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pyjwt>=2.8.0
watchdog>=3.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.82.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Monitoring and logging
structlog>=23.1.0
prometheus-client>=0.17.0
psutil>=5.9.0

# Optional: For advanced features
schedule>=1.2.0
click>=8.1.0